✅ [test] 실계좌 인증 삭제 테스트 코드 작성

+++ src/main/java/com/investmetic/domain/accountverification/service/AccountVerificationService.java
@@ -15,6 +15,7 @@ import com.investmetic.global.util.s3.FilePath;
 import com.investmetic.global.util.s3.S3FileService;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
@@ -81,14 +82,24 @@ public class AccountVerificationService {
     public void deleteStrategyAccountImages(Long strategyId, List<Long> requestDtoList) {
         for (Long requestDtoId : requestDtoList) {
             Optional<AccountVerification> targetImage = accountVerificationRepository.findById(requestDtoId);
+
+            // 삭제 요청한 이미지가 존재하지 않을 경우
             if (targetImage.isEmpty()) {
                 throw new BusinessException(ErrorCode.ACCOUNT_IMAGE_NOT_FOUND);
             }
+
+            // 해당 실계좌 인증 이미지의 전략아이디가 요청한 전략아이디와 일치하지 않을경우
+            if (!Objects.equals(targetImage.get().getStrategy().getStrategyId(), strategyId)) {
+                throw new BusinessException(ErrorCode.ENTITY_NOT_FOUND);
+            }
+
             AccountVerification imageEntity = targetImage.get();
+
             // FIXME: User 구현 후 수정 예정
 //            if (imageEntity.getCreatedBy() != user) {
 //                throw new BusinessException(ErrorCode.FORBIDDEN_ACCESS);
 //            }
+            
             accountVerificationRepository.delete(imageEntity);
         }
     }
